Expanded items and __init__ files
 Added a bunch of new item subclasses and expanded __init__ appropriately.
I thought currency could be a dummy class so we can just treat it differently when picking it up from a tile. When currency is picked up, rather than being added to the inventory, it simply adds the currency item's value to the player's 'currency reserve' (a new attribute). Having a dummy class will allow us to check if an item is currency or something else when picking up items.

 The others are self explanatory.