Shopping Update
Added a shop tile with functionality to buy and sell items. Player now has currency_amount attribute and the take inventory function was changed to show the amount of currency a player has. Also added the ability to drop items. The combat room is now a subclass of loot rooms to tidy code. The __init__ file was updated appropriately.